updated for 0.2.0 branch changes.

+++ src/tcs/log.c
@@ -30,7 +30,7 @@ void LogBlobData( char* szDescriptor, unsigned long sizeOfBlob, unsigned char* b
 {
 	char temp[1024];
 	char oneByte[8];
-	unsigned long i;
+	unsigned int i;
 
 	openlog(szDescriptor, LOG_NDELAY|LOG_PID, LOG_LOCAL5);
 	memset( temp, 0, sizeof( temp ));
